backend/pyproject.toml
[tool.poetry]
name = "todo-tracker-backend"
version = "0.1.0"
description = ""
authors = ["rodion <rpadakov@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
passlib = "^1.7.4"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg = "^3.2.3"
pytest-asyncio = "^0.24.0"
pyjwt = "^2.9.0"
pytest = "^8.3.3"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
